requirements.txt
# CORE
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# GOOGLE CLOUD
google-cloud-bigquery==3.14.1
google-cloud-storage==2.14.0
db-dtypes==1.2.0

# DATA & ML
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
xgboost==2.0.3
joblib==1.3.2

# LLM / AGENT
langchain==0.1.4
langchain-openai==0.0.5
openai==1.10.0

# API
fastapi==0.109.0
uvicorn[standard]==0.27.0

# UI
streamlit==1.31.0
plotly==5.18.0
requests==2.31.0

# TESTING
pytest==7.4.4
pytest-cov==4.1.0
httpx==0.26.0
ruff==0.1.14
